+++ src/runtime/RuntimeHelper.php
@@ -2,6 +2,7 @@
 namespace VovanVE\HtmlTemplate\runtime;
 
 use VovanVE\HtmlTemplate\components\ComponentInterface;
+use VovanVE\HtmlTemplate\components\ComponentSpawnerInterface;
 use VovanVE\HtmlTemplate\ConfigException;
 
 class RuntimeHelper implements RuntimeHelperInterface
@@ -9,7 +10,7 @@ class RuntimeHelper implements RuntimeHelperInterface
     /** @var array */
     private $params;
     /**
-     * @var array
+     * @var string[]|ComponentInterface[]|ComponentSpawnerInterface[]
      * @since 0.1.0
      */
     private $components = [];
@@ -128,17 +129,34 @@ class RuntimeHelper implements RuntimeHelperInterface
      */
     public function createComponent(string $name, array $properties = [], ?array $content = null): string
     {
-        /** @var string $component_class */
-        $component_class = $this->components[$name] ?? null;
-        if (null === $component_class) {
+        /** @var string $component_class_ */
+        $component_definition = $this->components[$name] ?? null;
+        if (null === $component_definition) {
             throw new ConfigException("Unknown component `$name`");
         }
-        if (!is_subclass_of($component_class, ComponentInterface::class)) {
-            throw new ConfigException("Component `$name` does not implement `ComponentInterface`");
+
+        if (is_string($component_definition)) {
+            if (!class_exists($component_definition)) {
+                throw new ConfigException("Component definition `$name` refers to unknown class");
+            }
+            if (!is_subclass_of($component_definition, ComponentInterface::class)) {
+                throw new ConfigException("Component `$name` does not implement `ComponentInterface`");
+            }
+
+            $component = new $component_definition($properties);
+        } elseif (is_object($component_definition)) {
+            if ($component_definition instanceof ComponentSpawnerInterface) {
+                $component = $component_definition->getComponent($properties);
+            } elseif ($component_definition instanceof ComponentInterface) {
+                $component = $component_definition;
+            } else {
+                throw new ConfigException("Component `$name` does not implement any of expected interfaces");
+            }
+        } else {
+            throw new ConfigException("Component definition `$name` has unsupported type");
         }
 
         /** @var ComponentInterface $component */
-        $component = new $component_class($properties);
         return $component->render($content);
     }