+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 HTML Template Changelog
 =======================
 
+0.1.1
+-----
+
+*   Add: Component definition now can be instance of:
+    *   `\VovanVE\HtmlTemplate\components\ComponentSpawnerInterface` to create
+        customized components at runtime;
+    *   `\VovanVE\HtmlTemplate\components\ComponentInterface` too render ready
+        component instance.
+
 0.1.0
 -----